Divide-into-packages
